pyproject.toml
[project]
name = "mufettis-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Habip ",email = "habipokcu9@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <4.0"
dependencies = [
    "pdfplumber (>=0.11.8,<0.12.0)",
    "chromadb (>=1.3.5,<2.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "fastapi (>=0.124.0,<0.125.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "google-genai (>=1.54.0,<2.0.0)",
    "docling (>=2.64.1,<3.0.0)",
    "transformers (>=4.57.3,<5.0.0)",
    "sentence-transformers (>=5.2.0,<6.0.0)",
    "rank-bm25 (>=0.2.2,<0.3.0)",
    "langchain (>=1.1.3,<2.0.0)",
    "langgraph (>=1.0.4,<2.0.0)",
    "langchain-google-genai (>=4.0.0,<5.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
